requirements/optional.txt
Pillow>=2.3.0
beautifulsoup4>=4.3.2
lxml>=3.3.1
pyinotify
